[Controller UI] Move cron scheduler + periodic tasks to Cluster Manager; trim sidebar label
The Quartz cron scheduler and the registered periodic-task names are
controller-level concerns (RetentionManager, SegmentRelocator, the
PinotTaskManager periodic that triggers minion task scheduling, etc.).
They are not minion task queues, so they belong on the Cluster Manager
home page rather than buried inside the Minion Task Manager page.

Changes
- New components/Homepage/CronSchedulerInformation.tsx wraps
  /tasks/scheduler/information into a collapsible accordion with
  scheduler name, instance id, version, thread pool size, jobs
  executed, running-since timestamp (timezone-aware), state, and
  scheduled job count. Tolerates both the legacy `getThreadPoolSize`
  controller key and a future `threadPoolSize` rename.
- New components/Homepage/PeriodicTasks.tsx wraps /periodictask/names
  into a collapsible accordion with a "Run Periodic Task" dialog that
  POSTs to /periodictask/run with optional table-name + table-type
  filters.
- HomePage.tsx mounts both new sections between Task Queues and
  Cluster Configuration.
- MinionTaskManager.tsx is trimmed to truly minion-only concerns:
  Task Types / Minion Instances / Running Tasks / Waiting Tasks tiles
  (Running/Waiting come from /tasks/summary) plus the Task Queues
  table. Cron scheduler accordion and periodic tasks section moved
  out.

Sidebar fix
- Renamed the sidebar entry from "Minion Task Manager" to "Minion
  Tasks". The longer label was wrapping to two lines in the open
  sidebar (icon above text), and the existing
  `.menu-item:hover { display: inline-block }` rule (intended for the
  collapsed-sidebar popup) un-wrapped it on hover, making the icon
  appear to jump. The shorter label fits on one line in both states
  and matches the user's original request ("add a tab of minion
  tasks on the left side bars"). The page heading remains
  "Minion Task Manager".

Verified
- tsc --noEmit clean for all touched files
- npm run build-dev succeeds
- ./mvnw license:check -pl pinot-controller passes
- Live UI walk on a quickstart-batch cluster:
  - Cluster Manager shows: Tenants, Instances, Task Queues,
    Cron Scheduler Information, Periodic Tasks (13), Cluster
    Configuration, Package Versions
  - Minion Tasks page shows: 4 minion-relevant tiles + Task Queues
  - Sidebar item height is 44px (single-line ListItem); icon no
    longer shifts on hover
  - Zero console errors across all routes
